Added type checks for security update
Line 24:

if !(user_id.is_a? Integer and team_id.is_a? Integer)
      flash[:error] = "Illegal parameter."

+++ app/models/teams_user.rb
@@ -21,8 +21,12 @@ class TeamsUser < ActiveRecord::Base
 
   # Removes entry in the TeamUsers table for the given user and given team id
   def self.remove_team(user_id, team_id)
-    team_user = TeamsUser.where(['user_id = ? and team_id = ?', user_id, team_id]).first
-    team_user.destroy unless team_user.nil?
+    if !(user_id.is_a? Integer and team_id.is_a? Integer)
+      flash[:error] = "Illegal parameter."
+    else
+      team_user = TeamsUser.where(['user_id = ? and team_id = ?', user_id, team_id]).first
+      team_user.destroy unless team_user.nil?
+    end
   end
 
   # Returns the first entry in the TeamUsers table for a given team id
